Fix mismatched schema from lit(0L) -> lit(0)

+++ mozetl/topline/topline_summary.py
@@ -215,7 +215,7 @@ def transform(dataframe, start, mode):
         .join(searches, attributes, "outer")
         .join(hours, attributes, "outer")
         .withColumnRenamed('country', 'geo')
-        .withColumn('crashes', F.lit(0))
+        .withColumn('crashes', F.lit(0).astype("long"))
         .na.fill(0)
     )
